+++ src/TinyGsmClientSIM7600.h
@@ -803,21 +803,21 @@ class TinyGsmSim7600
                                                 // TODO(?) Can 1 be returned
 
     // init variables
-    float ilat = 0;
-    char  north;
-    float ilon = 0;
-    char  east;
-    float ispeed       = 0;
-    float ialt         = 0;
-    int   ivsat        = 0;
-    int   iusat        = 0;
-    float iaccuracy    = 0;
-    int   iyear        = 0;
-    int   imonth       = 0;
-    int   iday         = 0;
-    int   ihour        = 0;
-    int   imin         = 0;
-    float secondWithSS = 0;
+    float   ilat = 0;
+    char    north;
+    float   ilon = 0;
+    char    east;
+    float   ispeed       = 0;
+    float   ialt         = 0;
+    int16_t ivsat        = 0;
+    int16_t iusat        = 0;
+    float   iaccuracy    = 0;
+    int16_t iyear        = 0;
+    int16_t imonth       = 0;
+    int16_t iday         = 0;
+    int16_t ihour        = 0;
+    int16_t imin         = 0;
+    float   secondWithSS = 0;
 
     streamSkipUntil(',');               // GPS satellite valid numbers
     streamSkipUntil(',');               // GLONASS satellite valid numbers
